feat: expose binary cache as a service
All interactions with the underlying kv and object stores for the binary cache has been abstracted behind a service.

The agent no longer requires permissions for the nar and nar-info stores, instead relying on a github.com/brianmcgee/nats.http to proxy http requests over NATS (as a transport) to the binary cache sitting behind the `nits.cache` subject.